"3pm. Just a bit more and I will resume. I am feeling a good amount of fear. But that is good. If posting a simple video is enough to put me this much in daze, think about how hard having to talk with people would suck.
3:25pm. Let me resume.

https://boards.4channel.org/biz/thread/53471429/is-this-just-modern-day-slavery-20-cut-on-your

///

Yes, is 20% but just for the first $500 you make, then is 10% until you reach $10,000, then is 5%. You only need two long-term clients and you are set, and if they like your work they will probably take you out of the platform.

> Designerfag making $150k in Upwork, created a network there and then got hired as a contractor in two Fortune 500 companies

///

Is this how it works? I wondered about that, and I guess this might be a route.

3:30pm. Let me start. Focus me, focus.

Let me start recording.

3:55pm. Done. This is really easy. With Camtasia, I can just record off the cuff, and then come in and edit out the pauses and taking of breaths.

4pm. Another masterpiece, let me go with it. It is also 1.5 minutes long.

4:05pm. https://youtu.be/9y4vkCoKgEI
Price comparison of Paperspace and competing cloud providers

Here it is.

4:10pm. Now let me make a video on how to start a gradient notebook.

4:20pm. Done recording the third video. Let me edit it.

Oh, almost forgot to shut down the machine.

5:35pm. An hour of editing for a 3.5 minutes video. But I am done. Let me render it.

5:45pm. Reviewed the video. It is good enough to post on Youtube.

6:25pm. Done with lunch. I've made a repo and uploaded the script to it.

6:35pm. I can't put in clickable links and have to mess with Google security.

6:50pm. I figured it out and now the video is up.

https://youtu.be/tapbr8kltyo
The easiest way to start a free tier notebook on Paperspace

Here it is.

6:55pm. I'll close here. I spent the whole day just to produce 6 minutes of video, but it is satisfying nonetheless. My problem is that I am speaking too quickly and without enough thought put into it. I can afford to have longer pauses between in order to load up the next few sentences properly. I need to learn to speak in a more determined manner. I need to put more thought into each of my words.

The way to achieve that is to recognize that video editing gives me time to catch my breath and think. I can take as long as I need between sentences. So that is the key.

Tomorrow I will do better at this. Today was ideal as an ice breaker though. Today I was far better than in the last two days.

7pm. I need to do better. Right now it is hardly better than doing a robo voiceover. I need to vividly visualize each of the sentences before firing them out. This will ease up the editing process afterwards. If I can do this I will be able nail those future interviews. Or so the theory goes.

Let me rest here.

7:10pm. It sure would be nice if this ends up being popular. I think I could make something promoting Paperspace on Youtube, as well as AI companies and such.

I can't imagine this sort of effort being less popular than Heaven's Key. That should be something to look forward to. Just pursue it. Pursue the passion, pursue the possibilities and pursue the dreams.

If I ever get the computational power that I desire, there will be time to make something of Simulacrum. If not, who cares. It will just be a possibility of the Singularity that never manifested. That I make money and the computational resources is more important than writing the story. That is the way to go."