"8:55pm. I figured out how the C# compilation works. It is simpler than it seems.
```cs
builder.RootComponents.Add<Qwe>("#appqwe");
```

Let me try this as my last experiment. And in

```html
<!DOCTYPE html>
<html lang="en">

<head>
    <meta charset="utf-8" />
    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
    <title>BlazorApp1</title>
    <base href="/" />
    <link href="css/bootstrap/bootstrap.min.css" rel="stylesheet" />
    <link href="css/app.css" rel="stylesheet" />
    <link href="BlazorApp1.styles.css" rel="stylesheet" />
</head>

<body>
    <div id="appqwe">Loading...</div>

    <div id="blazor-error-ui">
        An unhandled error has occurred.
        <a href="" class="reload">Reload</a>
        <a class="dismiss">🗙</a>
    </div>
    <script src="_framework/blazor.webassembly.js"></script>
</body>

</html>
```

I wonder if this is how it works.

...Yep. I understand it. I finally understand the structure of a Blazor project!

Let me go back a bit to the old thing.

9:25pm. Yeah, the old ASP.NET `.chtml` stuff is a wreck. What was MS thinking?

9:40pm. ASP.NET MVC is a pile of shit. No wonder I had trouble understanding it!

It takes quite a lot more to express than the new Blazor can. And in the latest .NET it seems it has been streamlined quite a bit. There has never been a better time to go into web development.

10pm. I am overdoing it. But I understand why I was confused back in 2020. Razor pages (.cshtml) I like a third of a reactive language. It buries me in the irrelevant, and has no power, so it is no wonder I got confused.

https://grpc.io/docs/what-is-grpc/introduction/

While looking at the docs I saw gRPC in ASP.NET, and decided to search for it on the JS. And what do I find, but this!

Yeah, this is something I could have used instead of ZeroMq.

It seems to have proper support than essentially being abandonware.

10:10am. Damn it, what am I going to do about Hopac. If anyone is going to bring it up to .NET Core 7 compability, that would have to be me. Hopefully it will work as is.

10:15pm. Well, I'll cross that bridge when I come to it.

At this point I am confused as to ASP.NET is. It is not Razor pages or Blazor. It is a...framework on which those two are built?

I can only interpret it like that.

https://github.com/giraffe-fsharp/Giraffe

https://dusted.codes/functional-aspnet-core

I think it is just about now that I can start to think about Giraffe.

> Ultimately I would like to build a web application in ASP.NET Core with a functional framework which does not only benefit from Kestrel but also from the entire ASP.NET Core eco system, including other middleware such as static files, authentication, authorization, security, the flexibility of the config system, logging or simply being able to retrieve information from the current hosting environment and more.

I do not really understand what this middleware is or what it offers me. I really have a while to go until I've made it as a webdev.

https://learn.microsoft.com/en-us/aspnet/core/fundamentals/static-files?view=aspnetcore-7.0

These fundamentals are something I haven't seen before.

This thing actually explains what that startup stuff in ASP.NET does!

> Think of it like this, ASP.NET Core is a web platform which sets the ground work for building any web application and MVC is a framework on top of the platform, which enables building web applications with an object oriented Model-View-Controller design pattern. Ideally as an F# developer I would like to replace the object oriented MVC framework with a functional equivalent, but keep the rest of ASP.NET Core's offering at the same time. This is very difficult with Suave at the moment.

This explains a lot. In order to really pass through to the door to the other side, I need to study these fundamentals.

> Kestrel is extremely fast and the Microsoft team is working hard on making it even faster.

I thought the Kestrel was something different than ASP.NET.

https://dusted.codes/functional-aspnet-core

This is a long article, I won't read it now. Let me close for real this time.

I am convinced now. If I want to understand ASP.NET, I should study Giraffe and then go from there. Though I need to do it as I study the fundamentals."