"8:20pm. https://www.goliquid.io/post/q-how-do-i-become-an-independent-contractor-in-croatia
> From registration to taxes and beyond, the info here will help you on your way to freelancing in Croatia.

I am looking up how to register as a contractor in Croatia. I asked my dad and he is super useless.

> A sole proprietor in Croatia is known as Obrt, and this status can be established under the Handicraft Act and the Companies Act. Obrt is the simplest form of business that a foreign professional can have in Croatia. The Obrt structure allows for growth, and once your company develops more, it is easy to change business structures,

https://useme.com/en/blog/freelancing-in-croatia-legal-requirements-and-tax-environment,147/#:~:text=Croatia%20offers%20a%20very%20good,between%202.1%20%25%20and%203.3%25.

Do freelancers in Croatia need to register or open a business?
> No, they do not need to. Freelancers can use the contract for specific work (ugovor o djelu) without opening a company. Furthermore, the tax system offers several ways for employers to work with freelancers. Freelancers usually use three types of contracts. One is the basic contract for specific work. The other is a contract with copyright transfer and a contract with an artist. With those three types of contracts, freelancers can work on different kinds of projects. The biggest difference between each contract is the statutory tax-deductible costs. We will provide more information in the following sections..

It seems I do not need to register an obrt.

> The most popular option for freelancers to pay the tax is via self-employment (paušalni obrt‘), as this option has simple tax rules. This option can be used for contractors who earn less than HRK 300k per year (Euro 40.3k). Moreover, a freelancer who has a full-time job does not need to take the full cost of the social contributions.

This is pretty low. I'll be making more than this as a programmer.

> Croatia offers a very good environment for freelancers with higher incomes. For freelancers with the self-employment option. Then they will pay extremely low-income tax. The tax rate is between 2.1 % and 3.3%. The only disadvantage is that there is no tax-free allowance for contracts for specific work and self-employment.

I do not understand this line.

///

What is the basic tax rate for freelancers in Croatia?
The tax rate is calculated on several features. For instance, the effective tax rate will be influenced by the copyright transfer with the artists. In general, the employer needs to pay 24% of the income tax. Below are the elements for the tax-deductible:

Basic contract for specific work
Contract with copyright transfer:30% of the gross amount
Contract with an artist:55% of the gross amount
We calculated to above data, the effective tax rate for different types of contracts are 31.6%, 22.13%, and 14.24%. This type of system is more favorable. Bulgaria and Poland also have this type of tax-deductible cost. In Poland, this type of tax-deductible cost is 50%.

///

https://contractortaxation.com/contracting-in-croatia/

No way around it, I'll have to ask around when the time comes.

https://www.reddit.com/r/croatia/comments/jcwqmw/freelancing_i_porez_kako_kada_koliko_za%C5%A1to/

> Bilo je hrpa dretvi na ovu temu, vec sam dvadeset puta napisao isti komentar valjda. Ukucaj u search sa strane “pausalni obrt”, “freelance” i slicno pa citaj. Takoder imas resurse u vidu Plavot ureda, Hrvatskog drustva nezavisnih profesionalaca i slicno. Ako ti nakon toga nesto nije jasno pitaj malo preciznije gdje si zapela, ovo tvoje pitanje je vrlo opsirno i jako ovisi o tvojoj situaciji, cime bi se bavila, na koji nacin, s kim itd.

https://plaviured.hr/
https://hdnp.hr/

///

Ugovor o djelu, isplata na žiro račun. Preko to dvoje porezna ima uvid u prihode i na miru si što se njih tiče, ne moraš raditi ni poreznu prijavu.

Ako se registriraš kao slobodni umjetnik, mislim da imaš povoljniji porez, pošto su umjetnici povlaštena klasa u HR, očito. :D

Ako dobivaš lovu izvana, onda porez i davanja plaćaš sama, ali ne znam točno kako to ide, da li plaćaš kad ti sjednu pare ili možeš sve odjednom u poreznoj prijavi. Pošalji mail poreznoj u svom gradu, objasnit će ti.

///

Kek. Here he is saying that artists are a priviledged class in Crotia.

https://www.iusinfo.hr/aktualno/u-sredistu/45240

This shit is pretty complicated.

https://www.youtube.com/results?search_query=ugovor+o+djelu+i+radu

https://youtu.be/Ua5n9wMSihM
Ugovori o radu i djelu, što se dobiva a što gubi?

9:50pm. This is the first time ever, I needed to think about taxes and my own finances. Ideally, I'd find some company outside the country that has very low taxes and have my income held by it. This would allow me to then transfer that money into a brokerage account and invest it without having it eaten by taxes.

I'll go that route.

https://www.justetf.com/en/etf-profile.html?groupField=none&sortField=sixMonthReturnEUR&sortOrder=desc&from=search&isin=DE0006289309#exposure

This is a really good site. Incidentally, if I were going now, this is what I would be buying. European banking stocks.

https://www.justetf.com/en/etf-profile.html?groupField=none&sortField=threeMonthReturnEUR&sortOrder=desc&from=search&isin=IE000DOZYQJ7#chart

You could make money shorting this.

https://www.justetf.com/en/etf-profile.html?groupField=none&sortField=sixMonthReturnEUR&sortOrder=desc&from=search&isin=IE00B1FZS574#chart

Turkey is a great investment. It has broken out particlarly strongly in the past half year, but now these earthquates are giving us an opportunity to come in cheaply.

Forget banks and go long Turkey. When you have a dead sector or country suddenly come to life like this, that is when you should pay attention. Right now Turkey is where it was in 2007.

Too bad I do not have cash to put to work.

The key thing to recognize is that speculative interest in it is certain to be dead at this point, meaning whatever is driving the valuations of this country stocks upwards are all strong hands.

https://www.justetf.com/en/find-etf.html?groupField=none&tab=rollreturn&sortField=sixMonthReturnEUR&sortOrder=desc&assetClass=class-bonds

It is good to look at bonds to get a sense of the sentiment.

https://www.justetf.com/en/etf-profile.html?groupField=none&sortField=sixMonthReturnEUR&sortOrder=desc&assetClass=class-bonds&from=search&isin=LU1109942653#chart

Yeah, at this point the bear market is dead according to the bond market. You could go long Turkey or euro banks, and if those don't work out in 1-3 months, switch to tech stocks or whatever. If you are wrong it is not like you will lose your shirt. Stocks aren't all or nothing. They should be pivoted in and out.

10:20pm. If I really want to be a proper investor, I need to make the smart choice once I get a job and not draw all of my pay where the taxman can get it. It is not like I need it to begin with.

The cool think about Croatia is that it has a 0% capital gains tax. It is a really good place to base yourself as a trader.

Forget that. I am wandering off into daydreams. Let's hope that I get something well paying. That is the first step. Who knows whether that will happen.

The only thing I can really do is continue learning and invest in my programming skills."