"2:15pm. Let me finally resume. I am into this simple course. It doesn't annoy me too much by backtracking and putting things that I am not aware of in.
2:30pm. https://youtu.be/G3e-cpL7ofc?t=3796
> In this lesson we are going to learn about one of the most useful tools for web development called the Chrome dev tools.

I think I'll take this chance to learn about the React tools as well as Chrome is not working as expected in the Inspector when it comes to React pages.

Actually it does work. I was wrong.

https://youtu.be/G3e-cpL7ofc?t=3881

Oh, I had no idea.

3:20pm. Let me take a break here.

3:40pm. I am back. Let me resume.

4:35pm. That break distrupted my pace, and now I find myself dozing off.

https://youtu.be/G3e-cpL7ofc?t=6258

Let me just say that if nothing else, this course teaching me how to use the Chrome dev tools is a huge help. I didn't know what CSS property is responsible for underlining, but it had it by default so all I needed to do was insect it and make use of that knowledge.

https://youtu.be/G3e-cpL7ofc?t=6444

Hmmm, but how would you turn this into a link.

Wait, wait. Am I getting fooled by the cursor being a pointer with span?

Nope, the link is not there.

https://youtu.be/G3e-cpL7ofc?t=6724

So he won't explain how to do actual links.

https://youtu.be/G3e-cpL7ofc?t=6843

So that is what the doctype is for.

https://youtu.be/G3e-cpL7ofc?t=7725

Again Google Fonts.

https://youtu.be/G3e-cpL7ofc?t=7876
> Now we are ready for the final project of this course - recreating Youtube.com.

https://youtu.be/G3e-cpL7ofc?t=8108

This is not good news to me. If img is a void elem, how am I going to put things on top of it?

6:30pm. Done with lunch.

Anyway, to answer my earlier questions, Google, I mean Youtube has some special tags and the image is only at the bottom. I do not know how that works yet.

...To think that one day I'd be studying Youtube HTML. I'll figure this mystery out later. let me watch a bit more. 10m more and then I'll probably close for the day.

https://youtu.be/G3e-cpL7ofc?t=8735

Interesting exercises.

6:50pm. https://youtu.be/G3e-cpL7ofc?t=8742

Let me stop here at the start of lesson 8.

I am 2:25h in out of 6:31h.

I wish I could study 24h a day, but at this point I am exhausted. I should be able to finish the html course in a day or two, and then I'll go back to the React one. After I deal with it, I should be set.

https://news.ycombinator.com/item?id=34869960
Running large language models like ChatGPT on a single GPU

Sigh...was I wrong to want to reach the Singularity my own way? Will the method in this link really work? Compresing the weights to 4 bits seems a lot. If what he is doing would really work, then MS and Goog would be all over it.

My biggest flaw is lack of money. I need to work on resolving it. It feels like I am never going to reach that goal, but I have to keep working on it.

7:05pm. I am going to get to a basic level of proficiency in Web UI work, just like I have on the desktop side and move on past this.

Once I have my porfolio pieces, I'll post them on the ML sub and see where that gets me.

I'll strive to make impression with my future moves. That is the path of a web developer. Nobody is will to support me so I have to hunt for attention. Attention equals money and power in today's economy. I need to finish what I've started in the first half of 2020 and master these two crucial skills. After that I will be ready to take on any kind of jobs.

Most ML jobs are essentially backend jobs plus using ML libraries on some occasion.

This time I will resolve myself to mastering all the hotshot technologies. Remember that time I applied to a ML place and had a SQL question that I made an error on? I am going to make sure that does not happen again.

The game will have a login, and proper registration like most online sites do. This will allow me to get familiar the basics of databases along the way.

Let me close here.

This is my usual programming pace like it has been for almost a decade now. I am back to who I was since the end of 2014. One day, may I change."