"https://youtu.be/PYCoRnJkn_c C# Developer Career Path Guide (ZERO TO HIRED)
This guy is trustworthy. He says that C# is a boring business language, and that I am going to be working on boring things with it. He does say that the salaries are huge.

https://youtu.be/PYCoRnJkn_c?t=232
> These days Javascript is mandatory in any development discipline.

https://youtu.be/PYCoRnJkn_c?t=464

Now he talks about building apps. He says the app should have all the features that one should see in a corporate website.

https://youtu.be/PYCoRnJkn_c?t=562
> Stay out of machine learning. Stay out of data science.

https://youtu.be/PYCoRnJkn_c?t=660
> Employers aren't going to call you if you don't have unit tests.

https://www.youtube.com/watch?v=PYCoRnJkn_c
> Download a couple of Udemy courses on how to actually interview.

7:15pm. Who is smart? Who is stupid here?

I had no idea there are Udemy courses on interviewing. I guess that is something to think about.

7:25pm. https://jobs.ashbyhq.com/motion/4f5f6a29-3af0-4d79-99a4-988ff7c5ba05?utm_source=hn

Look at jobs like these. Just why couldn't I get them? No reason at all.

One last note, he mentions I should have a cool hobby like running and that this is one of the reasons why employers are crazy about him. Egh, nevermind that. Getting out of the house was only ever on the schedule once I get out of the house. But that is not going to happen. I am on the path of self destruction.

https://www.udemy.com/course/find-a-job-interview-skills-training-course/

Oh this is a free course.

https://youtu.be/7ZBiSyl9f_E?t=105
> Half these jobs do not even have applicants.

8pm. https://youtu.be/BDMh2evivzw
Self-Taught C# Developer Road Map 2022

Let me watch this as well.

https://youtu.be/BDMh2evivzw?t=158
> You can do backend if you want to, but I highly, highly recommend that you do full stack.
> Because it is going to be way easier to get a job.

https://youtu.be/BDMh2evivzw?t=186
> C# is notoriously difficult to learn.

Really? He keeps saying that, but I do not believe him.

https://youtu.be/BDMh2evivzw?t=310
> There is no Python jobs where I live.

https://youtu.be/BDMh2evivzw?t=481
> Don't spend too much time, learn the basics, move on...to C#.

https://youtu.be/BDMh2evivzw?t=639

If you think about how long this would take a beginner, I am making progress at a blistering speed.

This is exactly the route I want to take, and he is pushing me in that direction. I really do want to add some knowledge about networking because I feel like an idiot when it comes to http requests and similar kinds of protocols.

Let me just take the time to do this properly. Taking an extra month to learn all of this and build up my portfolio.

https://youtu.be/BDMh2evivzw?t=785
> Focus Entitiy framework, stay away from MongoDB, and stay away from any type of fancy database.

https://youtu.be/BDMh2evivzw?t=907
> If you can build an app like that, you don't need a college degree, you don't even need any work history.

Huh, really? Ok, I'll give this a shot. I thought I might need to mess on the resume, but if a good app or two is enough.

https://youtu.be/BDMh2evivzw?t=929
> Is it going to be really difficult to make that app? Yes, but if...

Bullshit. This will be piss easy.

What this guy is saying is a good reality check for me. Right now I am feeling cursed, but if all I need is a good web app to get a good job, then that is a really easy goal to meet.

I am basically crazy right now. I feel like nothing is going right, that no matter what effort I make I can't meet my goals. That even if I do all this, I won't be able to achieve anything. But that can't be it.

I know that thanks to all the effort I put in, my programming skill should be remarkable.

https://youtu.be/BDMh2evivzw?t=1054
> You don't need to learn CSS all that much. You can get away with the Bootstrap framework.

https://youtu.be/BDMh2evivzw?t=1072
> I have a whole entire video on how to create a resume.

I am going to check it out in the future.

https://youtu.be/BDMh2evivzw?t=1155
> After the first year is when you start making crazy amounts of money.

He says that the first job is going to be crappy.

8:35pm. Ok, let's leave it at this. I'll follow the path outlined here. I am not going to be applying to any more jobs until I have everything set. I'll get my portfolio done, get the webdev skills that I need.

I won't pin too much hopes on the Valora F# job. That kind of depends on whether those people recognize my ability simply on the strength of the Spiral language project. And if my previous rounds are any indication, nobody cares.

I'll go the classical route as it will be so easy for me. Right now, I am already OP in terms of programming skills, I just need to reach out a bit and acquire the skills people are willing to pay for.

http://pythonnet.github.io/

> Python.NET (pythonnet) is a package that gives Python programmers nearly seamless integration with .NET Framework, .NET Core and Mono runtime on Windows, Linux and macOS. Python.NET provides a powerful application scripting tool for .NET developers. Using this package you can script .NET applications or build entire applications in Python, using .NET services and components written in any language that targets the CLR (C#, VB.NET, F#, C++/CLI).

> Python.NET is currently compatible and tested with Python releases 3.7 - 3.11.

Ohhh, it is up to day. That could be really cool.

I really need a way of transfering data between .NET and Python for my Holdem project later. Though I could do it over the network card, this way could be better.

I'll keep it in mind. I really want access to Python's ML libraries from my .NET backends, but not much else. I just need a way of transfering primitive arrays without the slowness of going through the network card. As long as I had that, I have everything I need for the Holdem project.

8:55pm. Let me close here for real. I need to believe that success and failure in the world make sense. I can't be paranoid and say that the black star is shining on me.

All I want is a chance to show off my skills and get a job based on that.

I can't really show off backend skills, so it makes sense to go the full stack route even if in the end all I get are backend jobs.

9pm. I should have achieved something with ML, but I literally can't do a thing with it right now that would be in any way impressive. The poker agents I'll dome is possibly the best one could do for somebody in my circumstances."