"8:15pm. https://hibox.live/elixir-for-humans-who-know-python
///

The first age of web development was static files served through a webserver. It was new, exciting, chaotic and fun, but very primitive. Most of you are probably too young to remember this era by now.

The second age of web development was in languages like Perl, PHP, Ruby and Python, where a client sends HTTP requests to URLs, and the server returns dynamic content rendered as a static resource. This was a very fun era to be a developer, but applications were limited in how complex and featureful they could be, and there were a fair inefficiencies from re-sending a whole page just to reflect a single updated value.

Live Apps combine the best of two worlds, enabling developers to build complex, dynamic applications without the need for any JavaScript or a dedicated API. A client requests an initial page load, which the server renders and maintains a local understanding of, and then makes a follow-up request to establish a WebSocket connection. Client side interactions with the application are then passed through the WebSocket and changes to the state are returned and the DOM is dynamically updated accordingly, without any API interaction or custom JavaScript required, as all of the logic is kept on the server-side.

///

Websockets?

9:20pm. https://desu-usergeneratedcontent.xyz/g/image/1677/26/1677261525409.jpg

///

>27
>neet
>no degrees
>never worked
>get help and dig myself out of depression
>start to learn coding because that's the only thing worth I can learn from home and it's fun.
Alternative is 3 year training by a school for retards and mentally ill to become a office clerk or gardener.
Maybe I should get my HS diploma and go to university, but that doesn't fix my problem of having literally years of blank space on my resume.

Should I just get really good at coding and fake my resume? Maybe it's truly おわりだ for me..

///

///

>>91740098
I asked Gigachad for you anon, remember that an literal AI will give you better answers than anyone on /g/. Now get the fuck out of my thread and start learning nigger.

///

Here is the image.

> Gigachad c.AI
> Get good enough at coding to show real results by creating a portfolio of projects. Then use those achievements to demonstrate your skills to potential employers without needing to use your resume - your work will speak for itself. This is the way, my brother.

Chatbots these days sure are great.

2/26/2023

9:40am. Let me check my mail.

9:45am. I've been getting some weird trolls lately.

Damn I am groggy. Let me read Kaiji and then I will get started on the SQL basics course. I need to figure out foreign keys. When Teddy explained them, they made sense, but then he started adding those collections and I am confused again.

Yesterday I got a grasp on how to connect to the SQL Server. I have verified that there is only a single server process that my client application connects to in the background.

Why was even something trivial like this confusing. Because I have like a dozens of old servers in the program list. Let me get rid of them.

9:50am. All cleaned up now. I only have the SQL Server 2022 stuff installed now.

https://mangadex.org/chapter/c0005ec5-40a2-4351-a7c5-24805dc4c074/24

Loooooooooooooooooollll. Old men are the best after all.

10am. Enough fun. Let me watch the SQL Server course.

Since I hadn't gotten any word from Valora at this time, can assume this is a dud? I don't know. I applied on the 23rd. 3 days is not too much, but if I do not get anything in another 1.5 weeks, I'll forget about it.

https://youtu.be/9Pzj7Aj25lw
Learn SQL in 1 Hour - SQL Basics for Beginners

Here is the one by Joey Blue. It is only 1h, so let me go for it.

https://youtu.be/9Pzj7Aj25lw?t=213

I see him having all these servers. I wonder how it is possible to rename it.

11:55am. Done with the course. It was pretty basic, but informative. This is the first time I've touched SQL. I find a syntax a bit confusing, but that is a matter of simply memorizing the relevant stuff. I can just Google it in the heat of battle. The biggest gain today and yesterday is my improved familiarity with SSMS. If I ever do SQL on a job, that should be convenient.

12pm. At this time I do not feel the need to learn more SQL.

What I really need to understand though is how EF and the...

Well, I suppose I could do a special record in SqlFun for joins, but I'd be better off just asking the author if it comes to that. It doesn't really matter right now. With ADO.NET if I go down that route, I can just use indexes to fetch everything back as a tuple.

12:05pm. This is a great time to stop and have breakfast.

Today I said that I had no email, but I wouldn't usually get anything on a Sunday anyway. I have no concepts of weekends in my personal life. I guess that will change when I get a job.

The next thing on the agenda is to get back to the Web API course and see it all the way through. I think I'l get a handle on foreign keys and relations in EF by the end of it.

I do not really care that much about SQL and databases, but I want to understand at least this much.

...Ah, come to think of it.

https://jacentino.github.io/SqlFun/Transforming-query-results

Hmmm, no inner joins here.

Nevermind that.

12:15pm. Let me commit here. There is no point in wondering about this. Just go through the EF course next and you are done with databases."