"8:15am. I am up. Somehow using games to push myself to brink leads to me falling into deep slumber. Today, my agenda is to finish CFR, but before that, I am overdue for a bath. Let me chill and then I will get to that.
https://www.reddit.com/r/MachineLearning/comments/11rizyb/d_anyone_else_witnessing_a_panic_inside_nlp_orgs/
> Let me tell you about quants : they do more math, they get more money and what do they do? They clean data to be fed into linear models.

9:40am. Done with the bath.

Let me get started.

10:35am. Let me resume. I gave myself the neti pot treatment and I'll see whether it makes a difference. Next time I'll do it more efficiently. I should let the salt dissolve in the glass after I microwaved it instead of putting in the neti pot and stirring later. That was foolish of me.

12:40pm. Let me pause here. So far I have 7:06m of the screecast recorded. Now I'll finally implement the update function for the CFR algorithm.

Maybe instead of fucking around and making these videos nobody will ever care to watch, I should just move on to interviewing.io instead? Let me jus finish this video, and then I will send out my faux resume to a couple of companies just to see what the response is.

I am not sure what I should be doing in this situation, but it is certainly not spending my entire days to make 7m of video. I'll finish this one though, definitely.

https://www.youtube.com/watch?v=7Pq-S557XQU
Humans Need Not Apply

1:25pm. Let me do the chores here. I want to resume quickly.

1:50pm. Let me resume.

2pm. Fuck Camtasia with a spoon. Why did it eat my most recently recorded video. Son a bitch.

...I realized it why. I really should be setting the option to display by date added. I rerecorded it, but once I found the setting, it showed me my old video near the top. Sigh. This happened once or twice before today, I think.

3pm. My god, my god! When I talk it builds up pressure inside my skull. And when I finish the sentence, the tension gets released and the there is a pop in my nasal pssage. The microphone picks it up every time. It is actually really difficult to blow some air at the end so that the pop happens later when I can cut it off.

Right now, this is my biggest enemy. I've gotten used to reciting, but this particular issue is just hard for me to deal with. Sometimes blowing air at the end works, but the click often happens before I can initiate it.

3:10pm. >  It would never learn anything in that situation.

I am trying to say this sentence without a pop at the end and it is killing me. It is not a matter of controlling pronounciation, it happens from somewhere inside my nasal passage once the pressure releases at the end.

https://www.sonnox.com/plugin/oxford-declicker

There are a ton of these plugins for this thing.

https://www.reddit.com/r/Reaper/comments/a4ol3w/how_to_achieve_a_midword_click_correction/

> Thanks! Luckily I nabbed iZotope de-click and de-noise during one of there sales. These two plugins save me literal hours, I don't want to remember life before them.

https://www.izotope.com/en/products/rx/features/de-click.html

It is on sale right now.

3:25pm. Instead of getting that, I am getting the advanced version from one of the warez sites I've stumbled upon back when I was trying to find Audiate. In hindsight, the reason why I couldn't find that one for free is most likely due to it being a cloud service. The pirates can't crack the TechSmith server. But this is different.

Had I known about it, I would have gotten Izotope Rx 10 Advanced, but I only found out about this program in the last 10 minutes. The only reason why I knew about the Izotope Declicker is because that Creators Village guy mentioned it. I looked for Izotope and here it is.

I guess it was a good thing I took a trip through the warezhouse. On IGetIntoPC there is only an order version. I found the most up to date version on DirtyWarez. Of course, you can't count on Google giving the good stuff to you these days. The best search engine would be Yandex at this date.

3:35pm. This distracted me. I'll put my faith in the program, but while it downloads, let me just continue recording in Camtasia. I'll stop putting so much effort in getting rid of the clicks manually and let the Rx take care of it later.

Having to constantly do this kind of editing is really killing both my pace and motivation.

3:40pm. Had to do a chore. Let me resume.

5:50pm. 13 full minutes today. It would really be much easier to do the script in its entirety. Task switching between talking and writing leads to delays.

5:55pm. I am not done with the video. What I need to do is introduce a new tab in the UI for training the agents. Something to act as a training board. I'll also look into charting here.

Right now, weirdly I am still not tired, but I did get up at 8am, so I should just take it easy and stop. The stress will go up easily enough once I cross a line. It is better to just be satisfied with a job well done.

If I want to do anything, why don't I check out a few Rx 10 tutorials and install the audio editor while I am at it? That would be for the best."